refactor the form page
