+++ components/Form/Select.styled.js
@@ -1,0 +1,23 @@
+const { styled } = require("styled-components");
+
+const Select = styled.select`
+  width: 100%;
+  padding: 1rem;
+  border-radius: 0.5rem;
+  border: 1px solid grey;
+
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+
+  &::after {
+    content: "↓";
+    position: absolute;
+    font-size: 1rem;
+    right: 1rem;
+    top: 0.95rem;
+    color: black;
+  }
+`;
+
+export default Select;
